Test for ForecastWeatherRepository complete.
Refactor: Create generic API parser.

Refactor the API error handling by introducing a generic `ApiParser`.
This change replaces the specific `LocationsApiParser` and `WeatherApiParser` with a single, unified implementation.

Dependencies for `turbine` are added and tests are updated to use it for testing flows. A `LogMockerExtension` for JUnit 5 is also added to simplify mocking `android.util.Log` in tests.